moved some helper functions to util file
